refactor: src/utils.js разделен на три файла в папке src/utils/
Функции по работе с маршрутом вынесены в класс Route (src/utils/Route.js),
где каждый метод сделан статичной функцией.
