+++ src/view/event-summary.js
@@ -1,5 +1,5 @@
-import {getDuration} from "../utils.js";
 import AbstractView from "./abstract.js";
+import {Route} from "../utils/route.js";
 
 export default class EventSummary extends AbstractView {
   constructor(event) {
@@ -23,7 +23,7 @@ export default class EventSummary extends AbstractView {
                 &mdash;
                 <time class="event__end-time" datetime="${endTime.toISOString()}">${endTime.toLocaleTimeString(`en-US`, {timeStyle: `short`, hour12: false})}</time>
               </p>
-              <p class="event__duration">${getDuration(beginTime, endTime)}</p>
+              <p class="event__duration">${Route.duration(beginTime, endTime)}</p>
             </div>
 
             <p class="event__price">
